+++ lumicks/pylake/nb_widgets/tests/test_kymotracker_widget.py
@@ -1,4 +1,6 @@
+from lumicks.pylake.kymotracker.detail.calibrated_images import CalibratedKymographChannel
 from lumicks.pylake.nb_widgets.kymotracker_widgets import KymoWidgetGreedy
+from lumicks.pylake.kymotracker.kymoline import KymoLine, KymoLineGroup
 from matplotlib.testing.decorators import cleanup
 import numpy as np
 import pytest
@@ -10,7 +12,7 @@ def test_widget_open(kymograph):
 
 
 @cleanup
-def test_parameters_kymo(kymograph, mockevent):
+def test_parameters_kymo(kymograph):
     """Test whether the parameter setting is passed correctly to the algorithm. By setting the threshold to different
     values we can check which lines are detected and use that to verify that the parameter is used."""
     kymo_widget = KymoWidgetGreedy(kymograph, "red", 1, use_widgets=False)
@@ -28,7 +30,7 @@ def test_parameters_kymo(kymograph, mockevent):
 
 
 @cleanup
-def test_invalid_algorithm_parameter(kymograph, mockevent):
+def test_invalid_algorithm_parameter(kymograph):
     kymo_widget = KymoWidgetGreedy(kymograph, "red", 1, use_widgets=False)
     with pytest.raises(TypeError):
         kymo_widget.algorithm_parameters["bob"] = 5
@@ -98,4 +100,39 @@ def test_refine_from_widget(kymograph, region_select):
     kymo_widget.refine()
     assert np.allclose(kymo_widget.lines[0].time_idx, np.arange(5, 20))
     assert np.allclose(kymo_widget.lines[0].coordinate_idx, [12] * 15)
+    assert len(kymo_widget.lines) == 1
+
+
+def test_stitch(kymograph, mockevent):
+    kymo_widget = KymoWidgetGreedy(kymograph, "red", 1, use_widgets=False)
+
+    k1 = KymoLine(
+        np.array([1, 2, 3]),
+        np.array([1, 1, 1]),
+        CalibratedKymographChannel.from_kymo(kymograph, "red"),
+    )
+    k2 = KymoLine(
+        np.array([6, 7, 8]),
+        np.array([3, 3, 3]),
+        CalibratedKymographChannel.from_kymo(kymograph, "red"),
+    )
+    kymo_widget.lines = KymoLineGroup([k1, k2])
+
+    # Go into line connection mode
+    kymo_widget._select_state({"new": "Connect Lines"})
+
+    assert len(kymo_widget.lines) == 2
+
+    # Drag but stop too early (not leading to a connected line)
+    kymo_widget._line_connector.button_down(mockevent(kymo_widget._axes, 3, 1, 3, 0))
+    kymo_widget._line_connector.button_release(mockevent(kymo_widget._axes, 4, 3, 3, 0))
+    assert len(kymo_widget.lines) == 2
+
+    # Drag all the way (stitch the two)
+    kymo_widget._line_connector.button_down(mockevent(kymo_widget._axes, 3, 1, 3, 0))
+    kymo_widget._line_connector.button_release(mockevent(kymo_widget._axes, 6, 3, 3, 0))
+
+    # Verify the stitched line
+    assert np.allclose(kymo_widget.lines[0].seconds, [1, 2, 3, 6, 7, 8])
+    assert np.allclose(kymo_widget.lines[0].position, [1, 1, 1, 3, 3, 3])
     assert len(kymo_widget.lines) == 1